+++ common/src/main/java/net/mehvahdjukaar/polytone/compat/EtfCompat.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.polytone.compat;
 
+import net.minecraft.client.renderer.entity.state.EntityRenderState;
 import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import traben.entity_texture_features.features.ETFManager;
 import traben.entity_texture_features.features.state.ETFEntityRenderState;
@@ -8,7 +9,7 @@ import java.util.UUID;
 
 public class EtfCompat {
 
-    public static int getLastKnownTextureVariantIndex(LivingEntityRenderState renderState) {
+    public static int getLastKnownTextureVariantIndex(EntityRenderState renderState) {
         UUID uuid = ((ETFEntityRenderState) renderState).uuid();
         return ETFManager.getInstance().LAST_SUFFIX_OF_ENTITY.getInt(uuid);
     }