+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/EntityRendererDispatcherMixin.java
@@ -1,0 +1,32 @@
+package net.mehvahdjukaar.polytone.mixins;
+
+import com.llamalad7.mixinextras.sugar.Local;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.polytone.Polytone;
+import net.minecraft.client.Camera;
+import net.minecraft.client.renderer.SubmitNodeCollector;
+import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
+import net.minecraft.client.renderer.entity.EntityRenderer;
+import net.minecraft.client.renderer.entity.state.EntityRenderState;
+import net.minecraft.client.renderer.state.CameraRenderState;
+import org.jspecify.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(EntityRenderDispatcher.class)
+public class EntityRendererDispatcherMixin {
+    @Shadow
+    @Nullable
+    public Camera camera;
+
+    @Inject(method = "submit", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/EntityRenderer;getRenderOffset(Lnet/minecraft/client/renderer/entity/state/EntityRenderState;)Lnet/minecraft/world/phys/Vec3;"))
+    private <S extends EntityRenderState> void polytone$onSubmit(
+            S entityRenderState, CameraRenderState cameraRenderState, double d, double e, double f,
+            PoseStack poseStack, SubmitNodeCollector submitNodeCollector, CallbackInfo ci,
+            @Local EntityRenderer<?,?> renderer) {
+        Polytone.ENTITY_MODIFIERS.captureRenderStates(cameraRenderState, renderer);
+    }
+}