+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/EntityRendererMixin.java
@@ -1,13 +1,20 @@
 package net.mehvahdjukaar.polytone.mixins;
 
 import com.llamalad7.mixinextras.sugar.Local;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.content.entity.IRenderStateWithId;
+import net.minecraft.client.renderer.SubmitNodeCollector;
+import net.minecraft.client.renderer.entity.BoatRenderer;
 import net.minecraft.client.renderer.entity.EntityRenderer;
+import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.entity.state.EntityRenderState;
+import net.minecraft.client.renderer.state.CameraRenderState;
 import net.minecraft.world.entity.Entity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(EntityRenderer.class)
@@ -17,4 +24,5 @@ public class EntityRendererMixin<T extends Entity, S extends EntityRenderState> 
     private void polytone$addId(T entity, float f, CallbackInfoReturnable<S> cir, @Local S renderState) {
         ((IRenderStateWithId) renderState).polytone$setId(entity.getId());
     }
+
 }