+++ fabric/src/main/java/net/mehvahdjukaar/polytone/integration/fabric/ModMenuCompat.java
@@ -1,0 +1,13 @@
+package net.mehvahdjukaar.polytone.integration.fabric;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import net.mehvahdjukaar.polytone.Polytone;
+
+public class ModMenuCompat implements ModMenuApi {
+
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return Polytone.CONFIGS::createScreenForMainMenu;
+    }
+}